+++ src/pycodedj/__init__.py
@@ -2,5 +2,5 @@
 
 from ._loop import loop, pattern
 
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 __all__ = ["loop", "pattern"]